# openapi.yaml
openapi: 3.0.3
info:
  title: Calorie Tracker API
  description: Backend API for the calorie tracker application (single-user MVP).
  version: 1.0.0
servers:
  - url: http://localhost:3010
    description: Local development
...
paths:
...
  /calorie-entry/calories-today:
    get:
      tags: [Calorie Entry]
      summary: Get total calories today
      description: Returns the sum of calories logged today for the current user. Requires Bearer token.
      operationId: getCaloriesToday
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: object
                    required: [totalCalories]
                    properties:
                      totalCalories:
                        type: number
                        format: float
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Use the token returned from POST /login as "Bearer &lt;token&gt;" (UUID, not JWT).
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error code
        message:
          type: string
          description: Human-readable message
...